Building AI Agents/2_phidata_Assisstant/requirements.txt
openai
phidata